+++ Docker/scripts_custardPy/plot_distance_map.py
@@ -5,6 +5,7 @@ import matplotlib.pyplot as plt
 import os
 import sys
 import numba
+
 # --------------------------------------------------------------------------------------------------
 # Set command line arguments
 argv = sys.argv